+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased 
 
+- perf: Scope sync to SentryCrash (#1193)
 - fix: Compile failure with C99 (#1224)
 
 ## 7.2.0-beta.4